refactor: Move login form to a separate component
